Implemented the library Managers accessors.
library manager encapsulates hashmaps. it processes loan requests by
searching the loan map for loans which contain the resource in the loan
request. if the resource is referenced by them the request will be
denied.

Refactored Record.java to be an Abstract class , instead of an
interface. Record.java provides a simple hash function based on javas
.hashCode() .