implemented gui and synchronous calls to LibraryManager
MainMenu.java acts as a wrapper class for the gui.
the LibManagerJob is a sub class of the swing worker class.
The LibManagerJob class is used together with actionlisteners in the
MainMenu class to communicate from the gui to the "database"
asynchronously.

All calls made to the LibraryManager by LibManagerJob class are
synchronized .